Update README.md
To build this app, you need to install Android Studio. Android Studio is the official IDE for building android apps. This is the link to download Android Studio including how to install it in your computer https://developer.android.com/studio/index.html

Figma was used to build or design the mockup. You can download figma by visiting this link https://www.figma.com/

When all is installed and set up, then we can start programming.

The Tic-Tac-Toe game can be played by two people or players and also by a single player and computer. In the single player mode, the first player is "X" while the second player is "O". Likewise in the multiplayer mode, the human player is "X" and the computer is "O".

The game has two different boards which are 3*3 and 5*5 respectively. The Single player 3*3 has different levels to choose from, but the Single player 5*5 only has one level. The reason for only one level is because there need to be an opportunity for the human player to win.

Copy the code to Android Studio or import it to Android Studio.

You can also get the APK in the file section. Unzip the APK, install in your android phone and enjoy the game.

Contributions are highly and respectfully welcome.